LKE-3931: Set maxVolumesPerNode to 7
This is a quick fix for the default case when we have 8 slots and one
slot is already taken by boot disk. There may be more slots in use by
non-PVC volumes such as swap but they are not the default in LKE.

Let's not let perfection be the enemy of good.