ci: Use go.mod for selecting the Go version
The "ci" workflow was previously using the "stable" and "oldstable"
versions, which as of Go 1.22 introduces a slight change in how go.mod
is formatted. This results in errors like

	go: updates to go.mod needed; to update it:
		go mod tidy

in the workflow logs.

So instead of relying on mutable version labels like "stable" and
"oldstable", this commit binds the version of Go used in CI to be what
is specified in the repository's go.mod file.

This will allow us the ability to gradually update Go versions without
needing to change the version in more than one place, and keep CI
predictable.