helm-chart: Get system serial number from DMI keyword
The previous invocation of dmidecode(8) relies on parsing some output
with awk(1). dmidecode(8) also provides the ability to return
information based on the DMI keyword for the table structure we are
interested in.

In this case, instead of looking for the System (type 1) DMI table, and
using awk(1) to pull out the system serial number, we can use

	dmidecode [-s|--string] system-serial-number

which will return the system serial number directly, and without the
need for any additional output parsing (except for stripping the
terminal newline character).