Add golangci-lint configuration
The golangci-lint configuration file contains some customized linter
configurations, simply because not all of the linters that are enabled
by default are worth it.

The "ci" workflow has also been modified a little. Namely, building the
binary is done first as a quick and inexpensive check. If it
successfully builds, then we go on to do the full-bore vet, staticcheck,
linting, and tests.