internal/driver: Fix off-by-one error with max volume attachments
When calculating the maximum number of allowed volume attachments, the
code was previously taking the ideal maximum number of volumes that
could be attached to a node. The way the attachment was calculated, it
treated instance disks the same as volumes, which is not correct.

This commit fixes what is effectively an off-by-one error, by
subtracting the number of instance disks from the theoretical maximum
number of block devices that can be attached to the instance.

In other words, controller and node servers will now report the number
of block storage volumes that can be attached, not just block devices.

Fixes #182