internal/driver: Rename and export the driver name
The driver's name was currently being stored in an unexported const:
driverName. This change renames that const symbol to "Name", which
results in it being exported from the package.

The driver name was also being redefined in main.go. By exporting the
internal/driver.Name const, it lets us use it from main.go.