internal/driver: Fix length check on volume label prefix
While the length check is a part of the regular expression used to
validate the optional volume label prefix, if the pattern does not match
the input string, then there is no way to indicate to the user how it
failed.

This commit breaks the check up into two, separate steps: checking the
length, then checking the character set.

When checking the length, we convert the string to a []rune first, since
we do not want to make any assumptions about the characters within the
string. Remember that in Go, strings must be valid UTF-8. If you iterate
over a string with a for-range loop, you will iterate over the
individual bytes which is incorrect in this case, since we do not know
if there are multi-byte characters in the given string. Converting the
string to a []rune lets us check the number of Unicode characters in the
string, and know if the string is too long.

Only after the length check is passed, will we look at each of the
individual characters.