internal/driver: Export WaitTimeout and CloneTimeout
Previously, these were unexported constants "waitTimeout" and
"cloneReadinessTimeout", respectively. waitTimeout was the number of
seconds to wait for most volume operations, and indicated how long we
should wait for the Linode API to tell us that a volume is "active" and
ready for use. cloneReadinessTimeout was again, the number of seconds to
wait for a volume to be "active", but only in the situation where a
volume was being created by cloning an existing volume.

These two constants have been converted to being time.Duration values,
and exported and documented with comments, since it would be good to be
able to know what those values are, if the need ever arises; they could
be looked up at
pkg.go.dev/github.com/linode/linode-block-storage-driver/internal/driver.WaitTimeout
for example.

While the changes to these two values are mostly for documentation, it
does add a tiny bit of additional handling when calling out to the
Linode API: the API expects the number of seconds to wait for a volume
to reach the desired state, so these time.Duration values need to be
converted to the number of seconds, as an int. That is where the new
waitTimeout() and cloneTimeout() functions come in.

Both of these functions -- waitTimeout, and cloneTimeout -- return the
number of seconds from WaitTimeout and CloneTimeout, respectively.