internal/driver: Export MinVolumeSizeBytes
MinVolumeSizeBytes is the minimum allowed size of a Linode block storage
volume, as of this commit's timestamp, in bytes.

The big silliness here is that the Linode API's volume endpoints return
sizes in "GB": not decimal gigabytes, not binary gibibytes, but some
funny interpretation of the number that looks like one, but is sometimes
calculated as the other.

In an effort to be consistent and less weird about disk sizes, this commit
defines the minimum size of a Linode block storage volume in *bytes*.
What this allows us to do, is not only be more accurate, but it makes
the math a lot simpler. Instead of doing potentially-expensive division
operations, we can go from bytes -> GB, and GB -> bytes, simply by
performing a right or left bitshift by 30:

	GB -> bytes		N << 30
	bytes -> GB		N >> 30

This commit also adds two convenience functions to help with doing these
conversions: gbToBytes and bytesToGB.