Fix - we need to return the FS path if the volume is attached to the linode
Reason - When the ControllerPublishVolume() is called and times out while waiting for volume to attach to the node, attacher will call the func again. Sometime, the volume might be attached by the time the func is re-executed. This change will check this case and return back the FS path so kubelet can use it.